+++ cpskin/core/behaviors/booking.py
@@ -1,39 +1,32 @@
 # -*- coding: utf-8 -*-
-from zope import schema
 
-from cpskin.locales import CPSkinMessageFactory as _
+from plone.autoform import directives
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.supermodel import model
+from zope import schema
 from zope.interface import provider
+
+from cpskin.locales import CPSkinMessageFactory as _
 
 
 @provider(IFormFieldProvider)
 class IBooking(model.Schema):
-    model.fieldset(
-        'bookingview',
-        label=_(
-            u'Booking'),
-        fields=(
-            'booking_enable',
-            'booking_price',
-            'booking_url',
-        ),
-    )
 
-    booking_enable = schema.Bool(
-        title=_(u'Booking'),
-        description=_(u'Is event booked'),  # noqa
+    directives.order_after(booking_type='IDublinCore.description')
+    booking_type = schema.Choice(
+        title=_(u'Booking type'),
         required=False,
-        default=False,
+        vocabulary=u'cpskin.core.vocabularies.booking_types',
     )
 
-    booking_price = schema.TextLine(
-        title=_(u'Tarriff'),
-        description=u'',
+    directives.order_after(booking_price='.booking_type')
+    booking_price = schema.Text(
+        title=_(u'Price'),
         required=False,
     )
 
+    directives.order_after(booking_url='.booking_price')
     booking_url = schema.URI(
-        title=_(u'Ticketing'),
+        title=_(u'Booking URL'),
         required=False,
     )